requirements.txt
# Core
fastapi==0.68.0
uvicorn[standard]==0.15.0
python-dotenv==0.19.0

# Database (using SQLite for development)
sqlalchemy==1.4.23
alembic==1.7.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5

# Utilities
httpx==0.19.0
email-validator==1.1.3
pydantic[email]==1.8.2

# Testing
pytest==6.2.5
pytest-cov==2.12.1
pytest-asyncio==0.15.1
